fixed some library issues.
all the models look nice and scale well, but the haptic rendering isn't working
on some of the models after the initial load.

can load the 5M point fireplace model and interact with it altough its a bit
slow at 100fps.

+++ pbrenderer/LYCudaHelper.cu
@@ -10,7 +10,7 @@ void LYCudaHelper::freeHostArray(void *devPtr)
 	checkCudaErrors(cudaFreeHost(devPtr));
 }
 
-void LYCudaHelper::allocateArray(void **devPtr, int size)
+void LYCudaHelper::allocateArray(void **devPtr, size_t size)
 {
 	checkCudaErrors(cudaMalloc(devPtr, size));
 }
