+++ pbrenderer/LYHapticDevice.h
@@ -64,10 +64,10 @@ public:
 
 	uint getProxyVBO() const { return m_ProxyObject->getVBO(); }
 	uint getProxyIB() const { return m_ProxyObject->getIB(); }
-	uint getProxyNumVertices() const { return m_ProxyObject->getNumVertices(); }
+	size_t getProxyNumVertices() const { return m_ProxyObject->getNumVertices(); }
 	uint getHIPVBO() const { return m_HIPObject->getVBO(); }
 	uint getHIPIB() const { return m_HIPObject->getIB(); }
-	uint getHIPNumVertices() const { return m_HIPObject->getNumVertices(); }
+	size_t getHIPNumVertices() const { return m_HIPObject->getNumVertices(); }
 	LYMesh* getHIPObject() const {return m_HIPObject;}
 	LYMesh* getProxyObject() const {return m_ProxyObject;}
